update version 1.4.2
/* Changelog v1.4.2
 *  add text shadow support
 *  change version utility system
 *  bug fixes
 *  code cleanup
 */

+++ src/main/java/com/nxtdelivery/quickStats/util/GetEntity.java
@@ -31,7 +31,7 @@ public class GetEntity {
 			if (mc.theWorld != null) {
 				mc.mcProfiler.startSection("pick");
 				mc.pointedEntity = null;
-				double d0 = (double) 200F; // this should be far enough
+				double d0 = 200F; // this should be far enough
 				mc.objectMouseOver = entity.rayTrace(d0, partialTicks);
 				double d1 = d0;
 				Vec3 vec3 = entity.getPositionEyes(partialTicks);
